Move dir for router
