+++ api/handler/router_test.go
@@ -1,4 +1,4 @@
-package server
+package handler
 
 import (
 	"bytes"
